requirements.txt
python-telegram-bot==20.6
schedule==1.2.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
icalendar==5.0.11
python-dateutil==2.9.0.post0
pyowm==3.3.0
python-dotenv==1.0.1
setuptools
feedparser